Clean up unit tests (#669)
